mcp_langflow_server/requirements.txt
mcp>=0.1.0
langflow-sdk>=0.1.0
fastapi>=0.68.0
uvicorn>=0.15.0